--- Dockerfile
@@ -1,12 +1,0 @@
-FROM mhart/alpine-node:10 as base
-WORKDIR /usr/src
-COPY package.json yarn.lock /usr/src/
-RUN yarn
-COPY . .
-RUN yarn run build && yarn --production
-
-FROM mhart/alpine-node:base-10
-WORKDIR /usr/src
-ENV NODE_ENV="production"
-COPY --from=base /usr/src .
-CMD ["node", "server.js"]
